create generate token in authRouter
