+++ package.json
@@ -20,6 +20,7 @@
 	},
 	"dependencies": {
 		"axios": "^0.19.0",
+		"bcrypt": "^3.0.6",
 		"bcryptjs": "^2.4.3",
 		"connect-session-knex": "^1.4.0",
 		"cors": "^2.8.5",
